requirements_updated.txt
# Retail Analytics System - Requirements
# Core dependencies for the enhanced retail tracking system

# Computer Vision and Deep Learning
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
opencv-python>=4.5.0

# Tracking and Detection
bytetrack>=1.0.0
numpy>=1.21.0
scipy>=1.7.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0

# Web Dashboard
flask>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0

# Image Processing
Pillow>=8.3.0
scikit-image>=0.18.0

# System and Utilities
tqdm>=4.62.0
pyyaml>=6.0
gdown>=4.0.0

# Development and Testing
pytest>=6.2.0
pytest-cov>=3.0.0
